[ng] moved date-adapter interface and abstract to /core - for tree shaking purposes
