+++ packages/ng/libraries/core/src/lib/date/date-adapter.interface.ts
@@ -1,0 +1,20 @@
+import { ELuDateGranularity } from './date-granularity.enum';
+
+export interface ILuDateAdapter<D> {
+	forge(year: number, month: number, date: number): D;
+	forgeToday(): D;
+	forgeInvalid(): D;
+	isValid(d: D): boolean;
+	compare(a: D, b: D, granularity: ELuDateGranularity): number;
+	isParsable(text: string): boolean;
+	parse(text: string): D;
+	format(d: D, format: string): string;
+	clone(d: D): D;
+
+	getYear(d: D): number;
+	getMonth(d: D): number;
+	getDate(d: D): number;
+	getDay(d: D): number;
+
+	add(d: D, count: number, granularity: ELuDateGranularity): D;
+}
