+++ packages/ng/libraries/core/src/lib/date/date-granularity.enum.ts
@@ -1,0 +1,6 @@
+export enum ELuDateGranularity {
+	day = 'day',
+	month = 'month',
+	year = 'year',
+	decade = 'decade',
+}
