+++ packages/ng/libraries/date/src/lib/adapter/date-adapter.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { ALuDateAdapter } from './date-adapter.class';
+import { ALuDateAdapter } from '@lucca-front/ng/core';
 
 @Pipe({
 	name: 'luDate',
