--- packages/ng/libraries/date/src/lib/adapter/date-granularity.enum.ts
@@ -1,6 +1,0 @@
-export enum DateGranularity {
-	day = 'day',
-	month = 'month',
-	year = 'year',
-	decade = 'decade',
-}
