+++ packages/ng/libraries/date/src/lib/adapter/index.ts
@@ -1,5 +1,2 @@
-export * from './date-granularity.enum';
-export * from './date-adapter.interface';
-export * from './date-adapter.class';
 export * from './date-adapter.module';
 export * from './date-adapter.pipe';
