+++ packages/ng/libraries/date/src/lib/native/native-date.adapter.ts
@@ -1,4 +1,4 @@
-import { ALuDateAdapter, ILuDateAdapter, DateGranularity } from '../adapter/index';
+import { ALuDateAdapter, ILuDateAdapter, ELuDateGranularity } from '@lucca-front/ng/core';
 import { LOCALE_ID, Inject, Injectable } from '@angular/core';
 import { getLocaleDateFormat, FormatWidth, formatDate } from '@angular/common';
 
@@ -98,21 +98,21 @@ export class LuNativeDateAdapter extends ALuDateAdapter<Date> implements ILuDate
 		return d.getDay();
 	}
 
-	add(d: Date, count: number, granularity: DateGranularity): Date {
+	add(d: Date, count: number, granularity: ELuDateGranularity): Date {
 		let year = this.getYear(d);
 		let month = this.getMonth(d);
 		let date = this.getDate(d);
 		switch (granularity) {
-			case DateGranularity.decade:
+			case ELuDateGranularity.decade:
 				year += 10 * count;
 				break;
-			case DateGranularity.year:
+			case ELuDateGranularity.year:
 				year += count;
 				break;
-			case DateGranularity.month:
+			case ELuDateGranularity.month:
 				month += count;
 				break;
-			case DateGranularity.day:
+			case ELuDateGranularity.day:
 				date += count;
 				break;
 		}
