+++ packages/ng/libraries/date/src/lib/native/native-date.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { ALuDateAdapter } from '../adapter/index';
+import { ALuDateAdapter } from '@lucca-front/ng/core';
 import { LuNativeDateAdapter } from './native-date.adapter';
 
 @NgModule({
